#sdy Fix bug where an in/out sharding with free+manual axes wasn't exported correctly.
Issue was that the `copy` and `FullToShard`/`ShardToFull` custom calls during `xla-sdy-mhlo-round-trip-shard-map-export`, the shardings we were creating when the shardings weren't fully manual were not accounting for the free/auto axes. It was just a fully open sharding attr accounting for any manual axes. Now we use the in/out sharding to account for the free axes.

PiperOrigin-RevId: 690673402