[xla:ffi] Relax type check on AnyBuffer.typed_data.
Currently we require the the requested type equal the underlying native buffer type, but I have a few cases where this is too strict (e.g. `std::complex != cuComplex`), so it would be useful to be able to cast to typed pointers of other types. In this change I've relaxed the check to only assert that the byte widths match.

PiperOrigin-RevId: 691741634